Some warnings removal and making CL code as default option

+++ examples/DeviceVector/deviceVector.cpp
@@ -52,7 +52,7 @@ int _tmain( int argc, _TCHAR* argv[ ] )
     pdV.reset( );
 
     std::cout << "Device Vector contents: " << std::endl;
-    for( size_t i = 0; i < vecSize; ++i )
+    for( int i = 0; i < vecSize; ++i )
     {
         std::cout << dV[ i ] << ", ";
     }
